refactor(types): hoist anchor_pair_is_consistent to ethlambda-types
State+Block invariant check belongs alongside the types it operates on,
not in storage. Storage, rpc test-driver, and the checkpoint-sync client
all import directly from ethlambda_types::state.
